requirements.txt
fastapi
pydantic
evolutionapi
pillow
uvicorn
python-multipart
opencv-python
numpy
mediapipe
ultralytics


